feat(world): bank deposit/withdraw (item+gold) + inventory tests
- Bank handler/service/serializer for OPENBANKWND, CLOSEBANKWND,
  PUTITEMBACK, GETITEMBACK, PUTGOLDBACK, GETGOLDBACK — account-shared
  tabs; mirrors DPSrvr.cpp OnPutItemBank/OnGetItemBank/OnPutGoldBank/OnGetGoldBank
- useItem: emit hp/mp/fp only when set
- item.schema: attack_speed is a fractional multiplier, not int
- tests: bank serializer, equip/consumable services, snapshot serializers

+++ packages/resources/src/schemas/item.schema.ts
@@ -86,8 +86,8 @@ export const ItemDefinitionSchema = z.object({
   /** Raw max attack (propItem dwAbilityMax) — weapons only */
   attack_max: z.number().int().min(0).optional(),
 
-  /** Raw attack speed (propItem dwAttackSpeed) — feeds combat atkSpeed table */
-  attack_speed: z.number().int().min(0).optional(),
+  /** Raw attack speed (propItem dwAttackSpeed) — fractional multiplier, feeds combat atkSpeed table */
+  attack_speed: z.number().min(0).optional(),
 
   /** Defense rating (armor only) */
   defense: z.number().int().min(0).optional(),
